+++ CHANGELOG.rst
@@ -4,7 +4,9 @@ Changelog
 
 Unreleased
 ----------
-* Add IPCC2006_PRIMAP categorization
+
+* Add IPCC2006_PRIMAP categorization.
+* Add refrigerant sub-classes and additional codes to CRFDI_class.
 
 0.5.1 (2021-05-04)
 ------------------